rfp-automation/requirements.txt
streamlit==1.49.1
plotly==6.3.0
pandas==2.3.2
numpy==2.3.3
python-dotenv==1.1.1
langsmith==0.4.29
